Clean up ALBEF example and add README
