+++ examples/albef/configs/retrieval.yaml
@@ -39,6 +39,7 @@ text_encoder_args:
   num_attention_heads: 12
   intermediate_size: 3072
   layer_norm_eps: 1e-12
+  dropout: 0.0
 
 multimodal_encoder_args:
   hidden_size: *hidden_size
