+++ examples/albef/finetune_vqa.py
@@ -10,23 +10,27 @@ import os
 import random
 import time
 
-import examples.utils.distributed as dist_utils
-
 import ruamel.yaml as yaml
 import torch
 import torch.backends.cudnn as cudnn
 import torch.distributed as dist
-from examples.albef.data.vqa_datamodules import VQADataModule
-from examples.albef.model import albef_model_for_vqa
-from examples.albef.utils import add_weight_decay
+from data.vqa_datamodules import VQADataModule
+from model import albef_model_for_vqa
 from torch.optim import AdamW
 from torch.optim.lr_scheduler import CosineAnnealingWarmRestarts
 
+from utils import (
+    add_weight_decay,
+    get_rank,
+    get_world_size,
+    init_distributed_mode,
+    is_dist_avail_and_initialized,
+    is_main_process,
+)
+
 
 def train(model, datamodule, args, device):
-    model_without_ddp = (
-        model.module if dist_utils.is_dist_avail_and_initialized() else model
-    )
+    model_without_ddp = model.module if is_dist_avail_and_initialized() else model
     model.train()
 
     optimizer_params = add_weight_decay(model, args["weight_decay"])
@@ -40,15 +44,15 @@ def train(model, datamodule, args, device):
     warmup_iterations = warmup_steps * step_size
 
     data_loader = datamodule.train_dataloader(
-        is_distributed=dist_utils.is_dist_avail_and_initialized(),
-        num_tasks=dist_utils.get_world_size(),
-        global_rank=dist_utils.get_rank(),
+        is_distributed=is_dist_avail_and_initialized(),
+        num_tasks=get_world_size(),
+        global_rank=get_rank(),
     )
 
     start_time = time.time()
 
     for epoch in range(args["max_epochs"]):
-        if dist_utils.is_dist_avail_and_initialized():
+        if is_dist_avail_and_initialized():
             data_loader.sampler.set_epoch(epoch)
 
         if epoch > 0:
@@ -104,7 +108,7 @@ def train(model, datamodule, args, device):
                 loss_str = "loss {}".format(loss.item())
                 print(time_str, epoch_str, batch_str, loss_str)
 
-        if dist_utils.is_main_process():
+        if is_main_process():
             save_obj = {
                 "model": model_without_ddp.state_dict(),
                 "optimizer": optimizer.state_dict(),
@@ -116,7 +120,7 @@ def train(model, datamodule, args, device):
                 os.path.join(args["checkpoint_root"], "vqa_checkpoint_%02d.pt" % epoch),
             )
 
-        if dist_utils.is_dist_avail_and_initialized():
+        if is_dist_avail_and_initialized():
             dist.barrier()
 
 
@@ -130,9 +134,9 @@ def evaluation(model, datamodule, args, device):
     answer_input_ids = datamodule.test_dataset.answer_input_ids.to(device)
     answer_atts = datamodule.test_dataset.answer_attention_mask.to(device)
     data_loader = datamodule.test_dataloader(
-        is_distributed=dist_utils.is_dist_avail_and_initialized(),
-        num_tasks=dist_utils.get_world_size(),
-        global_rank=dist_utils.get_rank(),
+        is_distributed=is_dist_avail_and_initialized(),
+        num_tasks=get_world_size(),
+        global_rank=get_rank(),
     )
 
     start_time = time.time()
@@ -174,10 +178,10 @@ def main():
     args = parser.parse_args()
     config = yaml.load(open(args.config, "r"), Loader=yaml.Loader)
 
-    dist_utils.init_distributed_mode(config)
+    init_distributed_mode(config)
     device = torch.device(config["device"])
 
-    seed = config["seed"] + dist_utils.get_rank()
+    seed = config["seed"] + get_rank()
     torch.manual_seed(seed)
     random.seed(seed)
     cudnn.benchmark = True
@@ -185,14 +189,14 @@ def main():
     datamodule = VQADataModule(**config["datamodule_args"])
     model = albef_model_for_vqa(config, pretrained=True)
     model = model.to(device)
-    if dist_utils.is_dist_avail_and_initialized():
+    if is_dist_avail_and_initialized():
         model = torch.nn.parallel.DistributedDataParallel(
             model, device_ids=[config["gpu"]]
         )
 
     train(model, datamodule, config["training_args"], device)
     result = evaluation(model, datamodule, config["eval_args"], device)
-    dist_utils.save_result(result, config["output_root"], "vqa_output")
+    save_result(result, config["output_root"], "vqa_output")
 
 
 if __name__ == "__main__":
