+++ examples/albef/requirements.txt
@@ -2,3 +2,4 @@ opencv-python==4.6.0.66
 pytorch-lightning==1.6.0
 Pillow==9.0.1
 ruamel_yaml==0.17.21
+transformers==4.24.0
